requirements.txt
Flask==3.0.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-cloud-aiplatform>=1.56.0
vertexai>=1.56.0

